moving RejectionTests to proper folder
