+++ src/System.ComponentModel.Composition/tests/System.ComponentModel.Composition.Tests.csproj
@@ -167,7 +167,6 @@
     <Compile Include="System\ComponentModel\Composition\UnitTesting\ExpectationOfIO.cs" />
     <Compile Include="System\ComponentModel\Composition\UnitTesting\ExportsAssert.cs" />
     <Compile Include="System\ComponentModel\Composition\UnitTesting\ExtendedAssert.cs" />
-    <Compile Include="System\ComponentModel\Composition\UnitTesting\RejectionTests.cs" />
     <Compile Include="System\ComponentModel\Composition\UnitTesting\RequiredCreationPolicyTests.cs" />
     <Compile Include="System\ComponentModel\Composition\UnitTesting\RetryMode.cs" />
     <Compile Include="System\ComponentModel\Composition\UnitTesting\TestServices.cs" />
@@ -182,6 +181,7 @@
     <Compile Include="System\Integration\ExportProviderEventTests.cs" />
     <Compile Include="System\Integration\LifetimeTests.cs" />
     <Compile Include="System\Integration\RecompositionTests.cs" />
+    <Compile Include="System\Integration\RejectionTests.cs" />
     <Compile Include="System\ReferenceTracker.cs" />
     <Compile Include="System\TypeExtensions.cs" />
     <Compile Include="TestAssembly.cs" />
